backend/requirements.txt
boto3==1.17.88
aiohttp==3.8.1
pydantic-argparse==0.5.0
requests==2.28.1
aiofiles==0.8.0
fastapi==0.79.0
mangum==0.15.1
uvicorn==0.18.2
aiohttp==3.8.1
gunicorn==20.0.4
SQLAlchemy==1.3.20
psycopg2==2.8.6
PyJWT==2.1.0
python-multipart==0.0.5
alembic==1.8.1